requirements.txt
customtkinter==5.2.2
faster-whisper==1.2.1
pyinstaller==6.17.0
Pillow==12.0.0
